updated performance numbers and text revisions

+++ D3337_Comparison/src/bgl_cc.hpp
@@ -9,4 +9,4 @@ G g;
 //populate g
 
 vector<size_t> c(num_vertices(g)); //components
-size_t num_cmps = connected_components(g, &c[0]);
+int num_cmps = connected_components(g, &c[0]);
